+++ src/pages/AccountPage.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const AccountPage = () => {
+  return (
+    <Container>
+      <h1>Account Page</h1>
+    </Container>
+  );
+};
+
+const Container = styled.div`
+  // CSS  
+`;
