+++ src/pages/LoginPage.jsx
@@ -1,7 +1,26 @@
+import { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 
 export const LoginPage = () => {
-  return <Container></Container>;
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    chrome.storage.local.get(["type"], (result) => {
+      if (result.type === "eezy") {
+        navigate("/eezy");
+      } else if (result.type === "squeeze") {
+        navigate("/squeeze");
+      }
+      chrome.storage.local.remove(["type"]);
+    });
+  }, [navigate]);
+
+  return (
+    <Container>
+      <h1>Login Page</h1>
+    </Container>
+  );
 };
 
 const Container = styled.div`
