+++ src/pages/SearchPage.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const SearchPage = () => {
+  return (
+    <Container>
+      <h1>Search Page</h1>
+    </Container>
+  );
+};
+
+const Container = styled.div`
+  // CSS
+`;
