+++ src/pages/UserSqueezyPage.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const UserSqueezyPage = () => {
+  return (
+    <Container>
+      <h1>User Squeezy Page</h1>
+    </Container>
+  );
+};
+
+const Container = styled.div`
+  // CSS
+`;
